% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{admin_build_pkgdown}
\alias{admin_build_pkgdown}
\title{Builds only the pkgdown site for the functions docs.}
\usage{
admin_build_pkgdown()
}
\value{
Nothing. Used for the side effect of building the pkgdown site.
}
\description{
Builds only the pkgdown site for the functions docs.
}
\keyword{internal}
