% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{admin_build_quarto}
\alias{admin_build_quarto}
\title{Builds only the Quarto website in the vignettes folder.}
\usage{
admin_build_quarto()
}
\value{
Nothing. Side effect of building the Quarto website.
}
\description{
Builds only the Quarto website in the vignettes folder.
}
\keyword{internal}
