% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{admin_create_project}
\alias{admin_create_project}
\title{Setup the project structure for working within the RAP.}
\usage{
admin_create_project(path)
}
\arguments{
\item{path}{File path and name of project.}
}
\value{
Nothing. Used for side effect of making project structure.
}
\description{
Setup the project structure for working within the RAP.
}
\examples{
\dontrun{
# Run proj_setup_git_config() first before running the admin_create_project() function.
# ukbAid::proj_setup_git_config("Luke W. Johnston", "lwjohnst@gmail.com")
admin_create_project(fs::path_temp("testproj"))
}
}
