% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-proposals.R
\name{admin_download_proposals}
\alias{admin_download_proposals}
\title{Download project proposals from Google Form.}
\usage{
admin_download_proposals(most_recent = FALSE)
}
\arguments{
\item{most_recent}{Whether to select only the most recent proposal.}
}
\value{
Invisibly returns data.frame. Used for side effect of saving
proposals as YAML file.
}
\description{
Download project proposals from Google Form.
}
