% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{admin_gh_add_contributor_to_repo}
\alias{admin_gh_add_contributor_to_repo}
\title{Add a contributor to a repo.}
\usage{
admin_gh_add_contributor_to_repo(user, proj_abbrev)
}
\arguments{
\item{user}{The GitHub username.}

\item{proj_abbrev}{The abbreviation for the project.}
}
\value{
Invisibly return results from GitHub API call as a list.
}
\description{
Add a contributor to a repo.
}
