% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{admin_gh_add_user_to_team}
\alias{admin_gh_add_user_to_team}
\title{Add the user to the SDCA GitHub Organization UK Biobank team.}
\usage{
admin_gh_add_user_to_team(user)
}
\arguments{
\item{user}{The GitHub username.}
}
\value{
Invisibly returns list of results of GitHub API calls. Used for the
side effect of adding user to GitHub.
}
\description{
Add the user to the SDCA GitHub Organization UK Biobank team.
}
