% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{admin_gh_list_repos}
\alias{admin_gh_list_repos}
\title{List all the repos on GitHub under the SDCA UK Biobank project.}
\usage{
admin_gh_list_repos()
}
\value{
A tibble.
}
\description{
List all the repos on GitHub under the SDCA UK Biobank project.
}
