% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{admin_list_approved_projects}
\alias{admin_list_approved_projects}
\title{List the approved projects connected to the UK Biobank at SDCA.}
\usage{
admin_list_approved_projects()
}
\value{
A tibble of project details.
}
\description{
List the approved projects connected to the UK Biobank at SDCA.
}
