% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{admin_read_projects}
\alias{admin_read_projects}
\title{Read all approved projects.}
\usage{
admin_read_projects()
}
\value{
A tibble of project details.
}
\description{
Show all approved projects, which are stored in YAML files, using the UK
Biobank at SDCA. Display this information as a \code{\link[tibble:tibble]{tibble::tibble()}}.
}
