% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{admin_start_approved_project}
\alias{admin_start_approved_project}
\title{Start and setup the approved project onto GitHub.}
\usage{
admin_start_approved_project(user, proj_abbrev)
}
\arguments{
\item{user}{The GitHub username.}

\item{proj_abbrev}{The abbreviation for the project.}
}
\value{
Nothing, used for the side effects of making a project and setting up
the GitHub repo.
}
\description{
Start and setup the approved project onto GitHub.
}
