% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-data.R
\name{builder_table_exporter}
\alias{builder_table_exporter}
\title{Build, but not run, the dx table exporter command.}
\usage{
builder_table_exporter(
  variables_to_extract,
  field = c("name", "title"),
  file_prefix = "data",
  project_id = get_rap_project_id(),
  dataset_record_id = dare_project_record_id,
  username = rap_get_user()
)
}
\arguments{
\item{variables_to_extract}{A character vector of variables you want to
extract from the UKB database (from the \code{.dataset} file). Use
\code{save_database_variables_to_project()} to create the
\code{data-raw/rap-variables.csv} file, open that file, and delete all variables you
don't want to keep. This is the file that contains the \code{variable_name} column
you would use for this argument.}

\item{file_prefix}{The prefix to add to the start of the file name. Defaults to "data".}

\item{project_id}{The project's abbreviation. Defaults to using
\code{get_rap_project_id()}, which is the name of the project folder.}

\item{dataset_record_id}{The "record ID" of the database, found when clicking
the \code{.dataset} file in the RAP project folder. Defaults to the ID for the
Steno project \code{dare_project_record_id}.}

\item{username}{The username to set where the dataset is saved to. Defaults to using
\code{rap_get_user()}, which is the name of the current user of the session.}
}
\value{
Outputs character string of command sent to \code{dx}.
}
\description{
This is mostly for testing purposes.
}
\examples{
library(tibble)
library(dplyr)
library(stringr)
library(magrittr)
rap_variables \%>\%
  sample_n(10) \%>\%
  pull(id) \%>\%
  builder_table_exporter(project_id = "test", username = "lwj") \%>\%
  cat()
}
