% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-data.R
\name{create_csv_from_database}
\alias{create_csv_from_database}
\title{Extract variables you want from the UKB database and create a CSV file to later upload to your own project.}
\usage{
create_csv_from_database(
  variables_to_extract,
  project_id = get_rap_project_id(),
  dataset_record_id = dare_project_record_id
)
}
\arguments{
\item{variables_to_extract}{A character vector of variables you want to
extract from the UKB database (from the \code{.dataset} file). Use
\code{save_database_variables_to_project()} to create the
\code{data-raw/rap-variables.csv} file, open that file, and delete all variables you
don't want to keep. This is the file that contains the \code{variable_name} column
you would use for this argument.}

\item{project_id}{The project's abbreviation. Defaults to using
\code{get_rap_project_id()}, which is the name of the project folder.}

\item{dataset_record_id}{The "record ID" of the database, found when clicking
the \code{.dataset} file in the RAP project folder. Defaults to the ID for the
Steno project \code{dare_project_record_id}.}
}
\value{
Outputs whether the extraction and creation of the data was
successful or not. Used for the side effect of creating the CSV on the RAP
server. The newly created CSV will have your username in the filename.
}
\description{
This function tells RAP to extract the variables you want from the \code{.dataset}
database file and to create a CSV file within the main RAP project folder.
When you want to use the CSV file in your own data analysis project, use
\code{download_project_data()}. You probably don't need to run this function often,
probably only once at the start of your project. NOTE: This function takes
5 minutes to run on the UKB RAP, so be careful with just randomly running it.
}
\examples{

\dontrun{
library(tidyverse)
read_csv("data-raw/rap-variables.csv") \%>\%
  pull(variable_name) \%>\%
  create_csv_from_database()
}
}
