% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-data.R
\name{download_data}
\alias{download_data}
\title{Downloads your UKB data file to the \code{data/data.csv} folder.}
\usage{
download_data(
  project_id = get_rap_project_id(),
  file_prefix = "data",
  file_ext = c("csv", "parquet"),
  username = rap_get_user()
)
}
\arguments{
\item{project_id}{The project's abbreviation. Defaults to using
\code{get_rap_project_id()}, which is the name of the project folder.}

\item{file_prefix}{The prefix to add to the start of the file name. Defaults to "data".}

\item{file_ext}{The file extension of the dataset. Defaults to "csv", but can also provide "parquet".}

\item{username}{The username to set where the dataset is saved to. Defaults to using
\code{rap_get_user()}, which is the name of the current user of the session.}
}
\value{
Downloads the data file to \verb{data/}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Use this function every time you open a new RStudio Session, aka every time you
use the UKB RAP. We don't want to save the data within the Git repository, so
you'd need to download it every time you go back to analyzing the data.
}
