% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-variables.R
\name{download_database_variables}
\alias{download_database_variables}
\title{Downloads the CSV file with the list of database variables and saves to \code{data-raw/rap-variables.csv}.}
\usage{
download_database_variables()
}
\value{
Doesn't return anything. Used for the side effect of downloading the data.
}
\description{
You would run this function when you just start your project, in order to have
a list of variables you want to use in your data analysis. You usually don't
need to do this more than once, unless you made a mistake in the file.
After downloading the variable list to \code{data-raw/rap-variables.csv}, you'd then
Git commit the downloaded file and afterwards would open it and remove
any variable you won't need.
}
