% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-data.R
\name{download_project_data}
\alias{download_project_data}
\title{Downloads your UKB CSV data file to the \code{data/data.csv} folder.}
\usage{
download_project_data(project_id = get_project_id(), username = get_username())
}
\arguments{
\item{project_id}{The project's abbreviation. Defaults to using
\code{get_project_id()}, which is the name of the project folder.}

\item{username}{A RAP username. Defaults to using \code{get_username()}.}
}
\value{
Downloads CSV to \code{data/data.csv}.
}
\description{
Use this function every time you open a new RStudio Session, aka every time you
use the UKB RAP. We don't want to save the data within the Git repository, so
you'd need to download it every time you go back to analyzing the data.
}
