% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj.R
\name{get_proj_dependencies}
\alias{get_proj_dependencies}
\title{Get package dependencies from the DESCRIPTION file.}
\usage{
get_proj_dependencies()
}
\value{
A character vector of package names.
}
\description{
Get package dependencies from the DESCRIPTION file.
}
