% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-data.R
\name{get_username}
\alias{get_username}
\title{Get the username of the user's who started the RStudio session.}
\usage{
get_username()
}
\value{
Character vector with users username within the RAP session.
}
\description{
Get the username of the user's who started the RStudio session.
}
