% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{gh_add_repo_to_team}
\alias{gh_add_repo_to_team}
\title{Add a GitHub repo to the UK Biobank Steno team.}
\usage{
gh_add_repo_to_team(repo)
}
\arguments{
\item{repo}{The name of the repository.}
}
\value{
Invisibly returns GitHub API response, used for side effect of
sending request to GitHub API.
}
\description{
Add a GitHub repo to the UK Biobank Steno team.
}
