% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{gh_add_user_to_repo}
\alias{gh_add_user_to_repo}
\title{Add a user to a repo.}
\usage{
gh_add_user_to_repo(user, repo)
}
\arguments{
\item{user}{The GitHub username.}
}
\value{
Invisibly returns NULL, used for side effect of sending request to GitHub API.
}
\description{
Add a user to a repo.
}
