% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{gh_add_user_to_team}
\alias{gh_add_user_to_team}
\title{Add the user to the UK Biobank GitHub team.}
\usage{
gh_add_user_to_team(user)
}
\arguments{
\item{user}{The GitHub username.}
}
\value{
Invisibly returns NULL. Used for the side effect of adding user to
GitHub.
}
\description{
Add the user to the UK Biobank GitHub team.
}
