% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{gh_create_repo}
\alias{gh_create_repo}
\title{Create a GitHub repository based on the existing local project.}
\usage{
gh_create_repo(path = ".")
}
\arguments{
\item{path}{Path to the local RStudio project. Defaults to current working
directory.}
}
\value{
The GitHub API response, used for the side effect of creating a
GitHub repository.
}
\description{
Create a GitHub repository based on the existing local project.
}
