% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{gh_create_tag}
\alias{gh_create_tag}
\title{Create a Git tag and push it to GitHub as a release.}
\usage{
gh_create_tag(version, message)
}
\arguments{
\item{version}{The version to set the tag (more specifically, the name of the
tag). Must be in the form of "##.##", like "1.0". A first number is for
submissions, and the second number is for fixes after the submission.}

\item{message}{The message to include when making the tag.}
}
\value{
Returns nothing, used for side effect of creating a tag on GitHub.
}
\description{
Create a Git tag and push it to GitHub as a release.
}
