% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{gh_get_repos}
\alias{gh_get_repos}
\title{Get all repo names in the UK Biobank GitHub team.}
\usage{
gh_get_repos()
}
\value{
A character vector.
}
\description{
Get all repo names in the UK Biobank GitHub team.
}
\examples{
\dontrun{
# Need to have a GitHub token set up.
gh_get_repos()
}
}
