% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{gh_get_team}
\alias{gh_get_team}
\title{Get the name of the Steno GitHub team for the UK Biobank project.}
\usage{
gh_get_team()
}
\value{
A character scalar.
}
\description{
Get the name of the Steno GitHub team for the UK Biobank project.
}
\examples{
gh_get_team()
}
