% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{gh_get_users}
\alias{gh_get_users}
\title{Get all users who are part of the UK Biobank Steno team on GitHub.}
\usage{
gh_get_users()
}
\value{
A character vector of usernames.
}
\description{
Get all users who are part of the UK Biobank Steno team on GitHub.
}
\examples{
\dontrun{
# Need to have a GitHub token set up.
gh_get_users()
}
}
