% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-variables.R
\name{git_commit_database_variables}
\alias{git_commit_database_variables}
\title{Commit the database variables to Git.}
\usage{
git_commit_database_variables()
}
\value{
Nothing. Adds and commits the \code{rap-variables.csv} file to Git.
}
\description{
Commit the database variables to Git.
}
