% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{git_commit_file}
\alias{git_commit_file}
\title{Add and commit a file to the Git history.}
\usage{
git_commit_file(file, message)
}
\arguments{
\item{file}{File to commit.}

\item{message}{Commit message.}
}
\value{
Nothing. Adds and commits a file.
}
\description{
Add and commit a file to the Git history.
}
\keyword{internal}
