% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj.R
\name{proj_create_path_dataset}
\alias{proj_create_path_dataset}
\title{Create a file name unique for a project and user, along with timestamp.}
\usage{
proj_create_path_dataset(user)
}
\arguments{
\item{user}{The name of the user you are creating this dataset for.}
}
\value{
A character scalar.
}
\description{
You use this when you want to create a dataset when using \code{\link[=proj_create_database]{proj_create_database()}}.
}
\examples{
\dontrun{
proj_create_path_dataset("lwjohnst")
}
}
