% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/proj.R
\name{get_proj_dependencies}
\alias{get_proj_dependencies}
\alias{proj_get_dependencies}
\title{Get package dependencies from the DESCRIPTION file.}
\usage{
get_proj_dependencies()

proj_get_dependencies()
}
\value{
A character vector of package names.
}
\description{
Get package dependencies from the DESCRIPTION file.
}
\section{Functions}{
\itemize{
\item \code{get_proj_dependencies()}: Deprecated function. Use \code{proj_get_dependencies()} instead.

}}
\examples{
proj_get_dependencies()
}
