% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj.R
\name{proj_get_git_config}
\alias{proj_get_git_config}
\title{Search Git's global config settings and list the user's name and email.}
\usage{
proj_get_git_config()
}
\value{
A tibble with the user's name and email.
}
\description{
Search Git's global config settings and list the user's name and email.
}
\examples{
proj_get_git_config()
}
