% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/proj.R
\name{setup_ukb_rap}
\alias{setup_ukb_rap}
\alias{proj_setup_rap}
\title{Run standard setup tasks in the UKB RAP after opening it up.}
\usage{
setup_ukb_rap()

proj_setup_rap()
}
\value{
Nothing. Used for it's side effects.
}
\description{
Run standard setup tasks in the UKB RAP after opening it up.
}
\section{Functions}{
\itemize{
\item \code{setup_ukb_rap()}: Deprecated function. Use \code{proj_setup_rap()} instead.

}}
