% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukbAid-package.R
\docType{data}
\name{project_variables}
\alias{project_variables}
\title{All variables found within the UK Biobank}
\format{
A data frame with 7663 rows and 3 variables:
\describe{
\item{id}{The ID for the "field" (or variable). Used to connect to the \code{rap_variables.csv} file.}
\item{ukb_variable_description}{Brief description of the variable.}
\item{link}{URL to the description of the variable.}
}
}
\source{
\url{https://biobank.ndph.ox.ac.uk/~bbdatan/Data_Dictionary_Showcase.tsv}
}
\usage{
project_variables
}
\description{
A dataframe that contains the IDs and variable descriptions for the variables
contained in the Steno DARE project on the UKB RAP.
}
\keyword{datasets}
