% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocol.R
\name{protocol_document}
\alias{protocol_document}
\title{Converts the protocol md or Rmd file into a Word document with relevant theme and settings.}
\usage{
protocol_document(toc = FALSE, toc_depth = 3, fig_width = 5, fig_height = 4)
}
\arguments{
\item{toc}{Include a table of contents?}

\item{toc_depth}{If a toc is included, how many heading levels should be shown?}

\item{fig_width}{The default width of the figures.}

\item{fig_height}{The default height of the figures.}
}
\value{
Outputs a Word document for the protocol.
}
\description{
Converts the protocol md or Rmd file into a Word document with relevant theme and settings.
}
