% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rap.R
\name{rap_copy}
\alias{rap_copy}
\alias{rap_copy_to}
\alias{rap_copy_from}
\title{Copying files between the RAP server and the RStudio RAP environment.}
\usage{
rap_copy_to(local_path, rap_path)

rap_copy_from(rap_path, local_path)
}
\arguments{
\item{local_path}{The path within the RStudio environment in the RAP.}

\item{rap_path}{The path within the RAP project server.}
}
\value{
The path of the copied file.
}
\description{
Copying files between the RAP server and the RStudio RAP environment.
}
\section{Functions}{
\itemize{
\item \code{rap_copy_to()}: Copying from the "local" RStudio environment on the RAP
and into the RAP project server.

\item \code{rap_copy_from()}: Copying from the RAP project server and into the "local"
RStudio environment on the RAP.

}}
