% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rap.R
\name{rap_get_database_entities}
\alias{rap_get_database_entities}
\title{Get a list of entities within the RAP database.}
\usage{
rap_get_database_entities()
}
\value{
A character vector of entities within the RAP database.
}
\description{
Get a list of entities within the RAP database.
}
\examples{
\dontrun{
rap_get_database_entities()
}
}
