% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rap.R
\name{rap_get_path_database}
\alias{rap_get_path_database}
\title{Get the path to the RAP UK Biobank dataset.}
\usage{
rap_get_path_database()
}
\value{
A path as a scalar character.
}
\description{
Get the path to the RAP UK Biobank dataset.
}
\examples{
\dontrun{
rap_get_path_database()
}
}
