% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rap.R
\name{rap_get_path_dirs}
\alias{rap_get_path_dirs}
\title{Get full paths of folders within the RAP project server.}
\usage{
rap_get_path_dirs(path)
}
\arguments{
\item{path}{The path that you want to see the folders within.}
}
\value{
A character vector.
}
\description{
Get full paths of folders within the RAP project server.
}
\examples{
\dontrun{
rap_get_path_dirs("users/")
}
}
