% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rap.R
\name{rap_get_path_schema}
\alias{rap_get_path_schema}
\title{Get the path to the RAP UK Biobank schema.}
\usage{
rap_get_path_schema()
}
\value{
A path as a scalar character.
}
\description{
Get the path to the RAP UK Biobank schema.
}
