% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rap.R
\name{rap_get_path_user}
\alias{rap_get_path_user}
\title{Get the path of a single user within the RAP project server.}
\usage{
rap_get_path_user(user)
}
\arguments{
\item{user}{A user's username.}
}
\value{
A character scalar of one users path.
}
\description{
Get the path of a single user within the RAP project server.
}
