% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rap.R
\name{rap_get_path_users}
\alias{rap_get_path_users}
\title{Get the paths of all the users within the RAP project server.}
\usage{
rap_get_path_users()
}
\value{
Character vector of folder paths.
}
\description{
Get the paths of all the users within the RAP project server.
}
\examples{
\dontrun{
rap_get_path_users()
}
}
