% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rap.R
\name{rap_get_project_id}
\alias{rap_get_project_id}
\title{Get the unique ID for the RAP project.}
\usage{
rap_get_project_id()
}
\value{
A character scalar.
}
\description{
Get the unique ID for the RAP project.
}
\examples{
\dontrun{
rap_get_project_id()
}
}
