% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rap.R
\name{rap_move_file}
\alias{rap_move_file}
\title{Move a file from one location to another within the RAP project server.}
\usage{
rap_move_file(path, output_dir)
}
\arguments{
\item{path}{The file path you want to move.}

\item{output_dir}{The folder (directory) you want to move the file to.}
}
\value{
The status results of the move command.
}
\description{
Move a file from one location to another within the RAP project server.
}
\examples{
\dontrun{
rap_get_path_files(".") |>
  stringr::str_subset("lwjohnst-.*\\\\.csv$") |>
  rap_move_file(
    rap_get_path_users() |>
      stringr::str_subset("lwjohnst")
  )
}
}
