% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukbAid-package.R
\docType{data}
\name{rap_variables}
\alias{rap_variables}
\title{Variable names that are specific to the RAP}
\format{
A data frame with 27,929 rows and 2 variables:
\describe{
\item{id}{The ID for the "field" (or variable) within the RAP UK Biobank dataset,
which represents the column in the dataset.}
\item{title}{The RAP variable title for the "field" (or variable). "Instance" is the
time point and "Array" is how the variable was split up with e.g. some
questionnaire variables. This is the human readable format.}
}
}
\usage{
rap_variables
}
\description{
A \code{\link[tibble:tibble]{tibble::tibble()}} that contains the field (variable) IDs and titles that are specific
to the RAP database and used to select from it.
}
\keyword{datasets}
