% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-data.R
\name{read_parquet}
\alias{read_parquet}
\title{Read a Parquet file and convert to DuckDB format.}
\usage{
read_parquet(path)
}
\arguments{
\item{path}{Path to the Parquet file.}
}
\value{
A SQL database object.
}
\description{
Read a Parquet file and convert to DuckDB format.
}
