% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-config.R
\name{setup_git_config}
\alias{setup_git_config}
\title{Add your name and email to Git's config settings.}
\usage{
setup_git_config(name, email)
}
\arguments{
\item{name}{Your full name, for instance "Luke W. Johnston".}

\item{email}{Your most commonly used email.}
}
\value{
Tibble with changed config settings. Also has side effect of
modifying the user.name and user.email settings in Git.
}
\description{
Add your name and email to Git's config settings.
}
