% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-config.R
\name{setup_rstudio_luke}
\alias{setup_rstudio_luke}
\title{Set basic configurations at the start of each new session in the UKB RAP.}
\usage{
setup_rstudio_luke()
}
\value{
NULL. The function is used for its side effects.
}
\description{
Set basic configurations at the start of each new session in the UKB RAP.
}
