% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-project.R
\name{setup_ukb_project}
\alias{setup_ukb_project}
\title{Setup the project structure for working within the RAP.}
\usage{
setup_ukb_project(path)
}
\arguments{
\item{path}{File path and name of project.}
}
\value{
Nothing. Used for side effect of making project structure.
}
\description{
Setup the project structure for working within the RAP.
}
\examples{

# Run setup_git_config() first before running the setup_ukb_project() function.
# ukbAid::setup_git_config("Luke W. Johnston", "lwjohnst@gmail.com")
setup_ukb_project(fs::path_temp("testproj"))

}
