% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-ukb-rap.R
\name{setup_ukb_rap}
\alias{setup_ukb_rap}
\title{Run standard setup tasks in the UKB RAP after opening it up.}
\usage{
setup_ukb_rap()
}
\value{
Nothing. Used for it's side effects.
}
\description{
Run standard setup tasks in the UKB RAP after opening it up.
}
