% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukbAid-package.R
\docType{data}
\name{statements}
\alias{statements}
\title{Textual statements required by UK Biobank.}
\format{
A list with 2 items:
\describe{
\item{protocol}{A text statement to include in the protocol.}
\item{paper}{A text statement to include in the paper.}
}
}
\usage{
statements
}
\description{
A list that contains two items, one to use in protocols and another to use
in a paper.
}
\keyword{datasets}
