% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-variables.R
\name{subset_rap_variables}
\alias{subset_rap_variables}
\title{Take the selected UKB variables from the protocol stage and subset the RAP specific variables list.}
\usage{
subset_rap_variables(
  project_variables_file = "data-raw/project-variables.csv",
  rap_variables_file = "data-raw/rap-variables.csv",
  save = TRUE
)
}
\arguments{
\item{project_variables_file}{The file location for the project variable CSV file. Defaults to \code{data-raw/project-variables.csv}.}

\item{rap_variables_file}{The file location for the RAP specific variable CSV file. Defaults to \code{data-raw/rap-variables.csv}.}

\item{save}{Whether to save (and overwrite) the newly subsetted RAP variables.}
}
\value{
A tibble of the newly subsetted RAP variable list. Save (overwrite) original RAP csv file.
}
\description{
This is done because RAP has a special naming system for their variables inside
the UK Biobank dataset that is slightly different from how the UK Biobank names them.
}
