% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukbAid-package.R
\docType{package}
\name{ukbAid-package}
\alias{ukbAid}
\alias{ukbAid-package}
\title{ukbAid: Aid Steno Researchers Who Work on the UKB RAP.}
\description{
Contains multiple functions that aid a researcher with setting up an analysis project on the UK Biobank Research Analysis Platform. Almost entirely designed around our needs within the Steno Aarhus Epidemiology Group.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/steno-aarhus/ukbAid}
  \item Report bugs at \url{https://github.com/steno-aarhus/ukbAid/issues}
}

}
\author{
\strong{Maintainer}: Luke Johnston \email{lwjohnst@gmail.com} (\href{https://orcid.org/0000-0003-4169-2616}{ORCID})

}
\keyword{internal}
