% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukbAid-package.R
\docType{data}
\name{variables}
\alias{variables}
\title{All variables found within the UK Biobank}
\format{
A data frame with 8272 rows and 3 variables:
\describe{
\item{field_id}{The ID for the "field" (or variable).}
\item{variable_name}{Full name of the variable.}
\item{link}{URL to the description of the variable.}
}
}
\source{
\url{https://biobank.ndph.ox.ac.uk/~bbdatan/Data_Dictionary_Showcase.tsv}
}
\usage{
variables
}
\description{
A dataframe that contains the field IDs and variable names for the over 8000
variables within the UK Biobank.
}
\keyword{datasets}
