% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zen.R
\name{zen_copy_json_template}
\alias{zen_copy_json_template}
\title{Copy the Zenodo metadata JSON template to the current working directory.}
\usage{
zen_copy_json_template()
}
\value{
Invisibly returns copied file path.
}
\description{
This JSON template contains the metadata needed when submitting a repo release
Zenodo.
}
