% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zen.R
\name{zen_create_file_record}
\alias{zen_create_file_record}
\title{Create a record for files on Zenodo.}
\usage{
zen_create_file_record(paths, metadata, token, sandbox = TRUE)
}
\arguments{
\item{paths}{The paths to the files to upload.}

\item{metadata}{The metadata necessary for Zenodo to create a deposit record.}

\item{token}{The token for Zenodo, via \code{\link[=zen_get_token]{zen_get_token()}}.}

\item{sandbox}{Whether to test the upload on the sandbox first. Strongly recommended.}
}
\value{
Nothing, used for the side effect of creating a record on Zenodo.
}
\description{
Create a record for files on Zenodo.
}
