% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zen.R
\name{zen_create_protocol_metadata}
\alias{zen_create_protocol_metadata}
\title{Create the metadata for Zenodo based on the Rmd or qmd protocol file.}
\usage{
zen_create_protocol_metadata(path)
}
\arguments{
\item{path}{The path to the protocol file.}
}
\value{
A list of metadata.
}
\description{
Create the metadata for Zenodo based on the Rmd or qmd protocol file.
}
\examples{
fs::path_package("ukbAid", "examples", "protocol.qmd") |>
  zen_create_protocol_metadata()
}
