% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zen.R
\name{zen_create_protocol_pdf}
\alias{zen_create_protocol_pdf}
\title{Create a PDF of the protocol Rmd or qmd file.}
\usage{
zen_create_protocol_pdf(path)
}
\arguments{
\item{path}{The path to the protocol file.}
}
\value{
Character vector of the created pdf file, used for side effect of
creating a PDF.
}
\description{
Create a PDF of the protocol Rmd or qmd file.
}
