% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zen.R
\name{zen_get_file_metadata_authors}
\alias{zen_get_file_metadata_authors}
\title{Extract author metadata from a Rmd or qmd file.}
\usage{
zen_get_file_metadata_authors(path)
}
\arguments{
\item{path}{The path to the file.}
}
\value{
A list with author details.
}
\description{
Extract author metadata from a Rmd or qmd file.
}
\examples{
fs::path_package("ukbAid", "examples", "protocol.qmd") |>
 zen_get_file_metadata_authors()
}
\keyword{internal}
