% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zen.R
\name{zen_get_project_github_url}
\alias{zen_get_project_github_url}
\title{Get the URL of the GitHub repository.}
\usage{
zen_get_project_github_url()
}
\value{
A list.
}
\description{
Get the URL of the GitHub repository.
}
\examples{
zen_get_project_github_url()
}
\keyword{internal}
