% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-data.R
\name{get_rap_project_id}
\alias{get_rap_project_id}
\title{Get the RAP project ID, will be the project main folder.}
\usage{
get_rap_project_id()
}
\value{
The project abbreviation/id, which is the name of the project folder
(and the same as the name of the \code{.Rproj} folder).
}
\description{
Get the RAP project ID, will be the project main folder.
}
