% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj.R
\name{proj_create_dataset}
\alias{proj_create_dataset}
\title{Extract the variables you want from the UKB database and create a CSV dataset.}
\usage{
proj_create_dataset(fields, output_path, database_id = dare_project_record_id)
}
\arguments{
\item{fields}{A character vector of field IDs you want to extract from the
UKB database. Select the variable ids from the \link{rap_variables} dataset.}

\item{output_path}{The name of the output CSV file.}

\item{database_id}{The ID of the dataset, defaults to the one we use in the
Steno UK Biobank DARE project.}
}
\value{
Outputs character string of the results from the command sent to RAP.
}
\description{
This function tells RAP to extract the variables you want from the \code{.dataset}
database file and to create a CSV file within the main RAP project folder.
When you want to use the CSV file in your own data analysis project, use
\code{download_project_data()}. You probably don't need to run this function often,
probably only once at the start of your project. NOTE: This function takes
5-60 minutes to run on the UKB RAP, so be careful with just randomly running it.
}
\examples{
\dontrun{
proj_create_dataset(
  c("eid", "p31", "p34"),
  proj_create_path_dataset(rap_get_user())
)
}
}
