% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj.R
\name{proj_get_dependencies}
\alias{proj_get_dependencies}
\alias{get_proj_dependencies}
\title{Get package dependencies from the DESCRIPTION file.}
\usage{
proj_get_dependencies()

get_proj_dependencies()
}
\value{
A character vector of package names.
}
\description{
Get package dependencies from the DESCRIPTION file.
}
\section{Functions}{
\itemize{
\item \code{get_proj_dependencies()}: Deprecated function. Use \code{proj_get_dependencies()} instead.

}}
