% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/proj.R
\name{get_rap_project_id}
\alias{get_rap_project_id}
\alias{proj_get_name}
\title{Get the RAP project ID, will be the project main folder.}
\usage{
get_rap_project_id()

proj_get_name()
}
\value{
The project abbreviation/id, which is the name of the project folder
(and the same as the name of the \code{.Rproj} folder).
}
\description{
Get the RAP project ID, will be the project main folder.
}
\section{Functions}{
\itemize{
\item \code{get_rap_project_id()}: Deprecated function. Use \code{proj_get_name()} instead.

}}
