% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rap.R
\name{rap_get_path_files}
\alias{rap_get_path_files}
\title{Get paths of files within a folder based on a given path in the RAP project server.}
\usage{
rap_get_path_files(path)
}
\arguments{
\item{path}{The path to search for files.}
}
\value{
A character vector.
}
\description{
Get paths of files within a folder based on a given path in the RAP project server.
}
\examples{
\dontrun{
rap_get_path_files(".")
}
}
