% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/rap.R
\name{get_username}
\alias{get_username}
\alias{rap_get_user}
\title{Get the username of the user who is working within the RAP RStudio session.}
\usage{
get_username()

rap_get_user()
}
\value{
Character vector with users username within the RAP session.
}
\description{
Get the username of the user who is working within the RAP RStudio session.
}
\section{Functions}{
\itemize{
\item \code{get_username()}: Deprecated function. Use \code{rap_get_user()} instead.

}}
\examples{
\dontrun{
rap_get_user()
}
}
