% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukbAid-package.R
\docType{data}
\name{rap_variables}
\alias{rap_variables}
\title{Variable names that are specific to the RAP}
\format{
A data frame with 22,355 rows and 3 variables:
\describe{
\item{id}{The ID for the "field" (or variable). Used when connecting with the \code{project_variables.csv} file.}
\item{field_id}{The RAP ID for the "field" (or variable). "Instance" is the
time point and "Array" is how the variable was split up with e.g. some
questionnaire variables.}
\item{rap_variable_name}{The name of the variable in the RAP database.}
}
}
\usage{
rap_variables
}
\description{
A dataframe that contains the field IDs and variable names that are specific
to the RAP database and used to select from it.
}
\keyword{datasets}
