% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{subset_rap_variables}
\alias{subset_rap_variables}
\title{Take the selected UKB variables from the protocol stage and subset the RAP specific variables list.}
\usage{
subset_rap_variables(
  project_variables_file = "data-raw/project-variables.csv",
  rap_variables_file = "data-raw/rap-variables.csv",
  instances = 0:9,
  save = TRUE
)
}
\arguments{
\item{project_variables_file}{The file location for the project variable CSV file. Defaults to \code{data-raw/project-variables.csv}.}

\item{rap_variables_file}{The file location for the RAP specific variable CSV file. Defaults to \code{data-raw/rap-variables.csv}.}

\item{instances}{An "instance" in UK Biobank is a time point of collection
for the variable. So a variable with Instance 0 and Instance 1 means that
the variable was collected twice at two different times. The default is to
select all the time points. Max is 9 instances, and while many variables
have up to 4 instances, the majority only have one (e.g. 0).}

\item{save}{Whether to save (and overwrite) the newly subsetted RAP variables.}
}
\value{
A tibble of the newly subsetted RAP variable list. Save (overwrite) original RAP csv file.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This is done because RAP has a special naming system for their variables inside
the UK Biobank dataset that is slightly different from how the UK Biobank names them.
}
