% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{upload_data}
\alias{upload_data}
\title{Upload a data file to the RAP.}
\usage{
upload_data(
  path,
  project_id = get_rap_project_id(),
  file_prefix = "data",
  username = rap_get_user()
)
}
\arguments{
\item{path}{Path to the dataset you want to upload.}

\item{project_id}{The project's abbreviation. Defaults to using
\code{get_rap_project_id()}, which is the name of the project folder.}

\item{file_prefix}{The prefix to add to the start of the file name. Defaults to "data".}

\item{username}{The username to set where the dataset is saved to. Defaults to using
\code{rap_get_user()}, which is the name of the current user of the session.}
}
\value{
A path to the uploaded data file.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
